"""fix""" repository, part. 1

+++ ministone/server/src/main/java/fr/ministone/repository/CardSpellRepository.java
@@ -7,7 +7,7 @@ import fr.ministone.game.card.CardSpell;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface CardSpellRepository extends CrudRepository<CardSpell, Integer> {
+public interface CardSpellRepository extends CrudRepository<CardSpell, Long> {
     CardSpell findByName(String name);
     Iterable<CardSpell> findAllByDeck(String deck);
 }
